Rename `txn_receipt.proto` -> `transaction_receipt.proto`
