+++ rest_api/tests/unit/test_receipt_requests.py
@@ -18,7 +18,7 @@ from aiohttp.test_utils import unittest_run_loop
 from components import Mocks, BaseApiTest
 from sawtooth_rest_api.protobuf.validator_pb2 import Message
 from sawtooth_rest_api.protobuf import client_receipt_pb2
-from sawtooth_rest_api.protobuf.txn_receipt_pb2 import TransactionReceipt
+from sawtooth_rest_api.protobuf.transaction_receipt_pb2 import TransactionReceipt
 
 
 class ReceiptGetRequestTests(BaseApiTest):
