+++ validator/tests/test_receipt_store/tests.py
@@ -28,8 +28,8 @@ from sawtooth_validator.networking.dispatch import HandlerStatus
 from sawtooth_validator.execution.tp_state_handlers import TpReceiptAddDataHandler
 
 from sawtooth_validator.protobuf import processor_pb2
-from sawtooth_validator.protobuf.txn_receipt_pb2 import TransactionReceipt
-from sawtooth_validator.protobuf.txn_receipt_pb2 import StateChange
+from sawtooth_validator.protobuf.transaction_receipt_pb2 import TransactionReceipt
+from sawtooth_validator.protobuf.transaction_receipt_pb2 import StateChange
 from sawtooth_validator.protobuf.client_receipt_pb2 import \
     ClientReceiptGetRequest
 from sawtooth_validator.protobuf.client_receipt_pb2 import \
